split test in 2. Added proper test for repr string. flake8
